requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/LLM - LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# AI/LLM - Agno
agno>=1.0.0

# OpenAI
openai>=1.10.0

# Redis
redis>=5.0.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Utils
tenacity>=8.2.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0
